chore(Events): Organize Events
