+++ src/main/java/net/refractored/bloodmoon/BloodmoonActuator.java
@@ -17,23 +17,15 @@ import org.bukkit.boss.BarFlag;
 import org.bukkit.boss.BarStyle;
 import org.bukkit.boss.BossBar;
 import org.bukkit.enchantments.Enchantment;
-import org.bukkit.entity.*;
-import org.bukkit.event.EventHandler;
-import org.bukkit.event.Listener;
-import org.bukkit.event.entity.EntityDamageByEntityEvent;
-import org.bukkit.event.entity.EntityDeathEvent;
-import org.bukkit.event.entity.PlayerDeathEvent;
-import org.bukkit.event.entity.SpawnerSpawnEvent;
-import org.bukkit.event.player.PlayerBedEnterEvent;
-import org.bukkit.event.player.PlayerJoinEvent;
-import org.bukkit.event.player.PlayerRespawnEvent;
-import org.bukkit.event.player.PlayerTeleportEvent;
+import org.bukkit.entity.EntityType;
+import org.bukkit.entity.LivingEntity;
+import org.bukkit.entity.Player;
+import org.bukkit.entity.SpawnCategory;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.meta.ItemMeta;
 import org.bukkit.metadata.MetadataValue;
 import org.bukkit.potion.PotionEffect;
 import org.bukkit.potion.PotionEffectType;
-import org.bukkit.projectiles.ProjectileSource;
 
 import java.io.Closeable;
 import java.util.*;
@@ -41,9 +33,9 @@ import java.util.*;
 /**
  * This is the class that handles most interaction during a BloodMoon
  */
-public class BloodmoonActuator implements Listener, Runnable, Closeable {
+public class BloodmoonActuator implements Runnable, Closeable {
     //Eligible mobs
-    public final EntityType[] rewardedTypes = {
+    public static final EntityType[] rewardedTypes = {
             EntityType.ZOMBIE,
             EntityType.SKELETON,
             EntityType.SPIDER,
@@ -60,14 +52,14 @@ public class BloodmoonActuator implements Listener, Runnable, Closeable {
 
     private static Map<World, BloodmoonActuator> actuators;
 
-    private World world;
-    private boolean inProgress;
+    public static World world;
+    private static boolean inProgress;
     private int bloodMoonLevel = 1;
-    private BossBar nightBar;
+    private static BossBar nightBar;
     private ActuatorPeriodic actuatorPeriodic;
 
-    private List<LivingEntity> blacklistedMobs;
-    private List<IBoss> bosses;
+    public static List<LivingEntity> blacklistedMobs;
+    public static List<IBoss> bosses;
 
     private void AddActuator (BloodmoonActuator instance) {
         if (actuators == null) actuators = new HashMap<>();
@@ -175,6 +167,13 @@ public class BloodmoonActuator implements Listener, Runnable, Closeable {
         }
         return eligiblePlayers;
     }
+    
+    private boolean isVanished(Player player) {
+        for (MetadataValue meta : player.getMetadata("vanished")) {
+            if (meta.asBoolean()) return true;
+        }
+        return false;
+    }
 
     public void SpawnHorde () {
         Random random = new Random();
@@ -183,12 +182,7 @@ public class BloodmoonActuator implements Listener, Runnable, Closeable {
         SpawnHorde(players.get(random.nextInt(players.size())));
     }
 
-    private boolean isVanished(Player player) {
-        for (MetadataValue meta : player.getMetadata("vanished")) {
-            if (meta.asBoolean()) return true;
-        }
-        return false;
-    }
+
 
     public void SpawnHorde (Player target) {
         if (target == null) return;
@@ -425,7 +419,7 @@ public class BloodmoonActuator implements Listener, Runnable, Closeable {
         nightBar = null;
     }
 
-    private void HideNightBarPlayer (Player player) {
+    public static void HideNightBarPlayer (Player player) {
         try
         {
             if (nightBar != null) nightBar.removePlayer(player);
@@ -449,7 +443,7 @@ public class BloodmoonActuator implements Listener, Runnable, Closeable {
         if (nightBar != null && percent >= 0.0 && percent <= 1.0f) nightBar.setProgress(1.0 - percent);
     }
 
-    private void HandleReconnectingPlayer (Player player) {
+    public static void HandleReconnectingPlayer (Player player) {
         if (isInProgress() && nightBar != null) nightBar.addPlayer(player);
         BroadcastBloodMoonWarningPlayer(player);
     }
@@ -462,7 +456,7 @@ public class BloodmoonActuator implements Listener, Runnable, Closeable {
     }
 
 
-    private void BroadcastBloodMoonWarningPlayer (Player player) {
+    private static void BroadcastBloodMoonWarningPlayer(Player player) {
         ConfigReader configReader = Bloodmoon.GetInstance().getConfigReader(world);
 
         LocaleReader.MessageLocale("BloodMoonWarningTitle", null, null, player);
@@ -483,7 +477,7 @@ public class BloodmoonActuator implements Listener, Runnable, Closeable {
      *
      * @return
      */
-    public ItemStack GetRandomBonus () {
+    public static ItemStack GetRandomBonus () {
 
         Random random = new Random(); //We want to regenerate it every time to ensure randomness
         Material itemMaterial;
@@ -556,7 +550,7 @@ public class BloodmoonActuator implements Listener, Runnable, Closeable {
         return null;
     }
 
-    public boolean IsInProtectedWGRegion (Player player){
+    public static boolean IsInProtectedWGRegion(Player player){
         try
         {
             RegionContainer rc = WorldGuard.getInstance().getPlatform().getRegionContainer();
@@ -573,7 +567,7 @@ public class BloodmoonActuator implements Listener, Runnable, Closeable {
         }
     }
 
-    private void ApplySpecialEffect (Player player, LivingEntity mob) {
+    public static void ApplySpecialEffect (Player player, LivingEntity mob) {
         if(IsInProtectedWGRegion(player)) return;
 
         ConfigReader configReader = Bloodmoon.GetInstance().getConfigReader(world);
@@ -610,257 +604,9 @@ public class BloodmoonActuator implements Listener, Runnable, Closeable {
         }
     }
 
-    public boolean isInProgress () {
+    public static boolean isInProgress() {
         ConfigReader reader = Bloodmoon.GetInstance().getConfigReader(world);
         return inProgress || reader.GetPermanentBloodMoonConfig();
-    }
-
-
-    //Events
-    @EventHandler
-    public void onPlayerConnect (PlayerJoinEvent event) {
-        if (isInProgress() && event.getPlayer().getWorld() == world)
-        {
-            HandleReconnectingPlayer(event.getPlayer());
-        }
-    }
-
-    @EventHandler
-    public void onPlayerTeleport (PlayerTeleportEvent event) {
-        World to = event.getTo().getWorld();
-        World from = event.getFrom().getWorld();
-        if (to != world && from != world) return; //None of our concern
-
-        if (from != to)
-        {
-            if (to == world && isInProgress())
-            {
-                //Someone entered our bm world
-                HandleReconnectingPlayer(event.getPlayer());
-            }
-            if (from == world && isInProgress())
-            {
-                //Someone left our bm world
-                HideNightBarPlayer(event.getPlayer());
-            }
-        }
-    }
-
-    @EventHandler
-    public void onPlayerRespawn (PlayerRespawnEvent event) {
-        World from = event.getPlayer().getWorld();
-        World to = event.getRespawnLocation().getWorld();
-        if (to != world && from != world) return; //None of our concern
-
-        if (from != to)
-        {
-            if (to == world && isInProgress())
-            {
-                //Someone respawned in our bm world
-                HandleReconnectingPlayer(event.getPlayer());
-            }
-            if (from == world)
-            {
-                //Someone respawned out of our bm world
-                HideNightBarPlayer(event.getPlayer());
-            }
-        }
-    }
-
-    @EventHandler
-    public void onPlayerDeath (PlayerDeathEvent event) {
-        if (!isInProgress()) return; //Only during BloodMoon
-
-        LocaleReader localeReader = Bloodmoon.GetInstance().getLocaleReader();
-        ConfigReader configReader = Bloodmoon.GetInstance().getConfigReader(world);
-
-        Player deadplayer = event.getEntity();
-        if (deadplayer.getWorld() != world) return; //Wrong world
-
-        if (configReader.GetLightningEffectConfig()) world.strikeLightningEffect(deadplayer.getLocation());
-
-        String deathMessage = event.getDeathMessage();
-
-        if (!deathMessage.contains(localeReader.GetLocaleString("DeathSuffix")))
-        {
-            deathMessage += " " + localeReader.GetLocaleString("DeathSuffix");
-
-            event.setDeathMessage(deathMessage);
-        }
-
-        if (configReader.GetExperienceLossConfig())
-        {
-            event.setNewTotalExp(0);
-            event.setDroppedExp(0);
-        }
-
-
-        if (configReader.GetInventoryLossConfig()) event.getDrops().clear();
-    }
-
-    @EventHandler
-    public void onPlayerSleeps (PlayerBedEnterEvent event) {
-        if (event.getPlayer().getWorld() == world)
-        {
-            if (isInProgress())
-            {
-                ConfigReader configReader = Bloodmoon.GetInstance().getConfigReader(world);
-                if (configReader.GetPreventSleepingConfig())
-                {
-                    LocaleReader localeReader = Bloodmoon.GetInstance().getLocaleReader();
-
-                    LocaleReader.MessageLocale("BedNotAllowed", null, null, event.getPlayer());
-                    event.setCancelled(true);
-                }
-            }
-        }
-    }
-
-    @EventHandler
-    public void onMobSpawn (SpawnerSpawnEvent event) {
-        ConfigReader configReader = Bloodmoon.GetInstance().getConfigReader(world);
-
-        if (configReader.GetMobsFromSpawnerNoRewardConfig() && event.getEntity().getWorld() == world && isInProgress())
-        {
-            for (EntityType type : rewardedTypes)
-            {
-                if (event.getEntityType() == type)
-                {
-                    if (event.getEntity() instanceof LivingEntity)
-                    {
-                        blacklistedMobs.add((LivingEntity) event.getEntity());
-                        break;
-                    }
-                }
-            }
-        }
-    }
-
-    @EventHandler
-    public void onMobDeath (EntityDeathEvent event) {
-        LivingEntity entity = event.getEntity();
-
-        for (IBoss boss : bosses)
-        {
-            if (entity == boss.GetHost())
-            {
-                Player killer = boss.GetHost().getKiller();
-                if (killer != null)
-                {
-
-                    LocaleReader.MessageAllLocale("BossSlain", new String[]{"$b", "$p"}, new String[]{boss.GetName(), killer.getName()}, world);
-                }
-
-                boss.Kill(killer != null && isInProgress());
-                bosses.remove(boss);
-                return;
-            }
-        }
-
-        if (!isInProgress()) return; //Only during BloodMoon
-
-        if (entity.getKiller() == null) return; //No killer, no reward
-
-        if (!entity.hasLineOfSight(entity.getKiller())) return; //No line of sight, no reward
-
-        if (event.getEntity() instanceof Player) return; //Handled in another method
-
-        ConfigReader configReader = Bloodmoon.GetInstance().getConfigReader(world);
-
-
-        if (entity.getWorld() != world) return; //Wrong world
-
-        if (blacklistedMobs.contains(entity))
-        {
-            //This mob was explicitly blacklisted. ignore it
-            blacklistedMobs.remove(entity);
-            return;
-        }
-
-        boolean eligible = false;
-        for (EntityType type : rewardedTypes)
-        {
-            if (entity.getType() == type) eligible = true;
-        }
-
-        if (!eligible) return; //Not eligible for reward
-
-        event.setDroppedExp(event.getDroppedExp() * configReader.GetExpMultConfig());
-
-        if (configReader.GetMobDeathThunderConfig())
-            world.strikeLightningEffect(event.getEntity().getLocation());
-
-        List<ItemStack> bonusDrops = new ArrayList<>();
-
-        int min = configReader.GetMinItemsDropConfig();
-        int max = configReader.GetMaxItemsDropConfig();
-
-        int itemCount = (max - min <= 0) ? min : new Random().nextInt(max - min) + min;
-
-        for (int i = 0; i < itemCount; i++)
-        {
-            bonusDrops.add(GetRandomBonus()); //Add the drops
-        }
-
-        for (ItemStack item : bonusDrops)
-        {
-            world.dropItemNaturally(entity.getLocation(), item); //Drop items
-        }
-    }
-
-    @EventHandler
-    public void onEntityDamage (EntityDamageByEntityEvent event) {
-        if (!isInProgress()) return; //Only during BloodMoon
-
-        Entity receiver = event.getEntity();
-        Entity damager = event.getDamager();
-
-        if (receiver.getWorld() != world || damager.getWorld() != world) return; //Wrong world
-        if (damager instanceof Projectile) //if its any damage dealing projectile
-        {
-            ProjectileSource source = ((Projectile) damager).getShooter(); //Get the shooter as Source
-            if (source instanceof LivingEntity) //Source is a mob, not a block
-            {
-                damager = (LivingEntity) source;
-            }
-        }
-
-
-        if (receiver instanceof LivingEntity && damager instanceof LivingEntity)
-        {
-            ConfigReader configReader = Bloodmoon.GetInstance().getConfigReader(world);
-            if (receiver instanceof Player)
-            {
-                for (EntityType type : rewardedTypes)
-                {
-                    if (damager.getType() == type)
-                    {
-                        //Player is damaged by monster
-                        if (event.getFinalDamage() == 0 && configReader.GetShieldPreventEffects())
-                            return; //Hit was shielded. We shall not apply configs
-                        event.setDamage(event.getDamage() * configReader.GetMobDamageMultConfig());
-                        ApplySpecialEffect((Player) receiver, (LivingEntity) damager);
-                        if (configReader.GetPlayerDamageSoundConfig())
-                            ((Player) receiver).playSound(receiver.getLocation(), Sound.AMBIENT_CAVE, 80.0f, 1.5f);
-                        if (configReader.GetPlayerHitParticleConfig())
-                            world.spawnParticle(Particle.FLAME, receiver.getLocation(), 60);
-                    }
-                }
-            } else if (damager instanceof Player)
-            {
-                for (EntityType type : rewardedTypes)
-                {
-                    if (receiver.getType() == type)
-                    {
-                        //Player dealt damage to monster
-                        event.setDamage((int) Math.ceil(event.getDamage() / configReader.GetMobHealthMultConfig()));
-                        if (configReader.GetMobHitParticleConfig())
-                            world.spawnParticle(Particle.CRIT_MAGIC, receiver.getLocation(), 60);
-
-                    }
-                }
-            }
-        }
     }
 
     /**
