+++ src/main/java/net/refractored/bloodmoon/commands/RegisterCommands.java
@@ -2,7 +2,8 @@ package net.refractored.bloodmoon.commands;
 
 import revxrsal.commands.bukkit.BukkitCommandHandler;
 
-public class RegisterCommands {
+public class
+RegisterCommands {
     public static void register(BukkitCommandHandler handler) {
         handler.register(new BloodmoonStart());
         handler.register(new BloodmoonStop());
