+++ src/main/resources/plugin.yml
@@ -7,43 +7,3 @@ softdepend:
 api-version: 1.20
 authors: [SpectralMemories, refractored]
 description: A BloodMoon shines every couple of nights
-#commands:
-#  bloodmoon:
-#    description: Base command. Should be followed by either show, start, stop or reload
-#    permission: "bloodmoon.bloodmoon"
-#    permission-message: "You do not have the permission to do this"
-#  #Uncomment this section to unlock the testsuite command. Official releases will obviously pass all tests, but if you want to
-#  #Run them yourself, knock yourself out!
-#  testsuite:
-#    description: Runs the integrated test suite. Usually only developers use this, there is however little harm to do it yourself. Test results will be in the BloodMoon folder
-#    permission: "bloodmoon.testsuite"
-#    permission-message: "You do not have the permission to do this"
-#permissions:
-#  bloodmoon.bloodmoon:
-#    description: Allows to run the base bloodmoon command
-#    default: true
-#  bloodmoon.show:
-#    description: Allows to see when a new bloodmoon will take place
-#    default: true
-#  bloodmoon.start:
-#    description: Allows to start a bloodmoon
-#    default: op
-#  bloodmoon.stop:
-#    description: Allows to end a bloodmoon
-#    default: op
-#  bloodmoon.reload:
-#    description: Allows to reload config files
-#    default: op
-#  bloodmoon.spawnzombieboss:
-#    description: Allows to spawn zombie boss at will
-#    default: op
-#  bloodmoon.killbosses:
-#    description: Allows to kill all bosses at will
-#    default: op
-#  bloodmoon.spawnhorde:
-#    description: Allows to spawn a horde
-#    default: op
-  #Required for the testsuite command
-  #bloodmoon.testsuite:
-  #  description: Allows to run the testsuite command. You most likely don't need this
-  #  default: op
